[Fix] Move ConfigHelper to AspNetFullFramework
