+++ src/Elastic.Apm.EntityFrameworkCore/EfCoreDiagnosticListener.cs
